+++ common/htlc_tx.h
@@ -8,7 +8,7 @@ struct preimage;
 struct pubkey;
 struct sha256_double;
 
-static inline u64 htlc_timeout_fee(u64 feerate_per_kw)
+static inline u64 htlc_timeout_fee(u32 feerate_per_kw)
 {
 	/* BOLT #3:
 	 *
@@ -17,10 +17,10 @@ static inline u64 htlc_timeout_fee(u64 feerate_per_kw)
 	 * 1. Multiply `feerate_per_kw` by 663 and divide by 1000 (rounding
 	 *    down).
 	 */
-	return feerate_per_kw * 663 / 1000;
+	return feerate_per_kw * 663ULL / 1000;
 }
 
-static inline u64 htlc_success_fee(u64 feerate_per_kw)
+static inline u64 htlc_success_fee(u32 feerate_per_kw)
 {
 	/* BOLT #3:
 	 *
@@ -29,7 +29,7 @@ static inline u64 htlc_success_fee(u64 feerate_per_kw)
 	 * 1. Multiply `feerate_per_kw` by 703 and divide by 1000 (rounding
 	 *    down).
 	 */
-	return feerate_per_kw * 703 / 1000;
+	return feerate_per_kw * 703ULL / 1000;
 }
 
 /* Create HTLC-success tx to spend a received HTLC commitment tx
@@ -39,7 +39,7 @@ struct bitcoin_tx *htlc_success_tx(const tal_t *ctx,
 				   unsigned int commit_output_number,
 				   u64 htlc_msatoshi,
 				   u16 to_self_delay,
-				   u64 feerate_per_kw,
+				   u32 feerate_per_kw,
 				   const struct keyset *keyset);
 
 /* Fill in the witness for HTLC-success tx produced above. */
@@ -60,7 +60,7 @@ struct bitcoin_tx *htlc_timeout_tx(const tal_t *ctx,
 				   u64 htlc_msatoshi,
 				   u32 cltv_expiry,
 				   u16 to_self_delay,
-				   u64 feerate_per_kw,
+				   u32 feerate_per_kw,
 				   const struct keyset *keyset);
 
 /* Fill in the witness for HTLC-timeout tx produced above. */
